requirements.txt
celery
coreschema
Django
django-celery-beat
django-timezone-field
djangorestframework
djangorestframework-simplejwt
drf-yasg
Jinja2
kavenegar
kombu
MarkupSafe
packaging
Pillow
prompt-toolkit
PyJWT
pyparsing
python-crontab
python-dateutil
python-decouple

